exits

+++ app/Http/Controllers/Scaffolder/ScaffolderController.php
@@ -3,16 +3,10 @@
 namespace App\Http\Controllers\Scaffolder;
 
 use App\Http\Controllers\Controller;
-
-
-
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\DB;
-
 use Illuminate\Support\Facades\File;
-
-use function foo\func;
 
 
 class ScaffolderController extends Controller
@@ -29,18 +23,18 @@ class ScaffolderController extends Controller
 
         $tables = DB::select(DB::raw("SELECT TABLE_NAME AS _table FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = '$request->db'"));
 
-        foreach ($tables as $key=>$table) {
+        foreach ($tables as $key => $table) {
             $t = $table->_table;
             $columns = DB::select(DB::raw("show fields from " . $t));
 
 
-            $atr=[];
+            $atr = [];
             foreach ($columns as $column) {
-                $f=$column->Field;
+                $f = $column->Field;
 
-                $atr[$f]=$column;
+                $atr[$f] = $column;
             }
-            $metadados[$t]=$atr;
+            $metadados[$t] = $atr;
 
         }
 
@@ -48,9 +42,10 @@ class ScaffolderController extends Controller
         return view("scaffolder.configuretables", compact("metadados"));
     }
 
-    public function tablesConfigureP1Post(Request $request){
+    public function tablesConfigureP1Post(Request $request)
+    {
 
-        $json=json_encode($request->except('_token'), JSON_PRETTY_PRINT);
+        $json = json_encode($request->except('_token'), JSON_PRETTY_PRINT);
 
         file_put_contents(base_path('app/Http/Controllers/Scaffolder/data/metadados.json'), stripslashes($json));
 
@@ -60,19 +55,18 @@ class ScaffolderController extends Controller
         dd("Criado tudo");
 
 
-
         dd($json);
     }
 
-    private function createByJsonObject($json){
+    private function createByJsonObject($json)
+    {
 
-        $json=collect(json_decode($json));
-        $json=$json->first();
+        $json = collect(json_decode($json));
+        $json = $json->first();
 
 
-        foreach($json as $m)
-        {
-            if($m->enable=="yes"){
+        foreach ($json as $m) {
+            if ($m->enable == "yes") {
 
 
                 Artisan::call("make:model $m->modelName   --controller");
@@ -82,47 +76,49 @@ class ScaffolderController extends Controller
             }
 
 
-
         }
     }
 
 
-    private function populateFiles($m){
+    private function populateFiles($m)
+    {
+
+        $modelPath = base_path("app/" . $m->modelName . ".php");
 
 
+        if (!File::exists($modelPath)) {
 
-        $modelPath=base_path("app/".$m->modelName.".php");
+            $contents = File::get($modelPath);
 
+            $contents = substr_replace($contents, "", -3);
+            $contents .= "\n";
+            $contents .= '    protected $table = "' . $m->modelTable . '";';
+            $contents .= "\n";
+            $contents .= '    protected $fillable=[';
+            $i = 0;
+            $len = collect($m->fields)->count();
 
-        $contents = File::get($modelPath);
-
-        $contents=substr_replace($contents ,"",-3);
-        $contents.="\n";
-        $contents.='    protected $table = "'.$m->modelTable.'";';
-        $contents.="\n";
-        $contents.='    protected $fillable=[';
-        $i =0;
-        $len = collect($m->fields)->count();
-
-            foreach ($m->fields as $key=>$field){
+            foreach ($m->fields as $key => $field) {
                 $i++;
 
-                if($i==$len){
+                if ($i == $len) {
 
 
-                    $contents.='"'.$key.'"';
-                }else{
-                    $contents.='"'.$key.'",';
+                    $contents .= '"' . $key . '"';
+                } else {
+                    $contents .= '"' . $key . '",';
                 }
 
             }
-        $contents.="];\n";
-        $contents.="}";
+            $contents .= "];\n";
+            $contents .= "}";
 
-        file_put_contents($modelPath, $contents);
+            file_put_contents($modelPath, $contents);
 
+        }else{
 
-
+            //pensar em alguma coisa
+        }
 
 
     }
